requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
boto3==1.34.83
apache-airflow-providers-amazon==8.13.0
python-dotenv==1.0.1
matplotlib==3.8.4
